+++ modules/logs/versions.tf
@@ -5,5 +5,13 @@ terraform {
       source  = "hashicorp/azurerm"
       version = ">= 2.1"
     }
+    external = {
+      source  = "hashicorp/external"
+      version = "~> 2"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = "~> 3"
+    }
   }
 }